requirements.txt
pyTelegraBotAPI
duckduckgo_search
python-dotenv